feat: Enhance models and configuration management - Optimize models module with unified precision management and output formatting - Add spin_flip symmetry support in configuration space, halving network inference cost - Improve global precision control with one-click GPU computation precision switching - Enhance timing functionality for better performance monitoring

+++ lever/workflow/fitter.py
@@ -123,7 +123,7 @@ class Fitter:
         feat_c = ctx.features_c
 
         # Preallocate energy buffer on device
-        energies_init = jnp.empty((max_steps,), dtype=jnp.float64)
+        energies_init = jnp.empty((max_steps,))
 
         # carry = (state, energies, k, last_energy, streak, converged)
         carry_init = (
